Network.JsonRpc.Server.Types

Description

This module provides support for writing handlers for JSON-RPC endpoints

Documentation

data MaybeContent a

The entire JSON RPC api is collapsed to a single endpoint. Therefore, we need a type that may or may not return content.

Constructors

SomeContent !a

EmptyContent

type MaybeJsonRpcResponse = MaybeContent (JsonRpcResponse Value Value)

type RawJsonRpcEndpoint = ReqBody '[JsonRpcContentType] (Request Value) :> Post '[JsonRpcContentType] MaybeJsonRpcResponse

class RouteJsonRpc a where

This internal class is how we accumulate a map of handlers for dispatch

Associated Types

type RpcHandler a (m :: Type -> Type)

Methods

jsonRpcRouter :: Monad m => Proxy a -> Proxy m -> RpcHandler a m -> Map String (Value -> m (MaybeContent (Either (JsonRpcErr Value) Value)))

hoistRpcRouter :: Proxy a -> (forall x. m x -> n x) -> RpcHandler a m -> RpcHandler a n

generalizeResponse :: (ToJSON e, ToJSON r) => Either (JsonRpcErr e) r -> Either (JsonRpcErr Value) Value

onDecodeFail :: IsString e => String -> JsonRpcErr e

serveJsonRpc :: (Monad m, RouteJsonRpc a) => Proxy a -> Proxy m -> RpcHandler a m -> Request Value -> m MaybeJsonRpcResponse

Collapse a to a single handler to handle RawJsonRpc

Orphan instances

(RouteJsonRpc api, HasContextEntry (context .++ DefaultErrorFormatters) ErrorFormatters) => HasServer (RawJsonRpc api :: Type) context

Instance details

Associated Types

type ServerT (RawJsonRpc api) m

Methods

route :: Proxy (RawJsonRpc api) -> Context context -> Delayed env (Server (RawJsonRpc api)) -> Router env

hoistServerWithContext :: Proxy (RawJsonRpc api) -> Proxy context -> (forall x. m x -> n x) -> ServerT (RawJsonRpc api) m -> ServerT (RawJsonRpc api) n