Network.JsonRpc.Types

Description

Work with JSON-RPC protocol

API specification types

data RawJsonRpc api

A JSON RPC server handles any number of methods.

data JsonRpc (method :: Symbol) p e r

JSON-RPC endpoint which respond with a result given a query.

The type has the following type parameters:

method: the name of the JSON-RPC method

p: the params field type in the JSON-RPC request

e: the error type for the result field in the JSON-RPC response

r: the value type of the result field in the JSON-RPC response

data JsonRpcNotification (method :: Symbol) p

JSON-RPC endpoints which do not respond

data JsonRpcContentType

The JSON-RPC content type.

JSON-RPC messages

data Request p

Client messages

Constructors

Request

Fields

method :: !String

params :: !p

requestId :: !(Maybe Word64)

omitted for notification type messages

data JsonRpcErr e

Constructors

JsonRpcErr

Fields

errorCode :: !Int

errorMessage :: !String

errorData :: !(Maybe e)

data JsonRpcResponse e r

Server Ack message

Constructors

Result !Word64 !r

Ack !Word64

Errors !(Maybe Word64) !(JsonRpcErr e)

data UnusedRequestParams

Constructors

UnusedRequestParams

Smart constructors for standard JSON-RPC errors

mkJsonRpcParseErr :: Maybe e -> JsonRpcErr e

JSON-RPC error codes based on JSONRPC Spec

mkJsonRpcInvalidRequestErr :: Maybe e -> JsonRpcErr e

mkJsonRpcMethodNotFoundErr :: Maybe e -> JsonRpcErr e

mkJsonRpcInvalidParamsErr :: Maybe e -> JsonRpcErr e

mkJsonRpcInternalErr :: Maybe e -> JsonRpcErr e

Type rewriting

type family JsonRpcEndpoint a where ...

JSON-RPC

Equations

JsonRpcEndpoint (JsonRpc m p e r) = ReqBody '[JsonRpcContentType] (Request p) :> Post '[JsonRpcContentType] (JsonRpcResponse e r)

JsonRpcEndpoint (JsonRpcNotification m p) = ReqBody '[JsonRpcContentType] (Request p) :> Post '[JsonRpcContentType] NoContent